+++ fltk/examples/counter3.rs
@@ -4,9 +4,9 @@ use fltk::{
     *,
 };
 
-const BLUE: Color = Color::from_rgbi(0x42A5F500);
-const SEL_BLUE: Color = Color::from_rgbi(0x2196F300);
-const GRAY: Color = Color::from_rgbi(0x75757500);
+const BLUE: Color = Color::from_hex(0x42A5F5);
+const SEL_BLUE: Color = Color::from_hex(0x2196F3);
+const GRAY: Color = Color::from_hex(0x757575);
 const WIDTH: i32 = 600;
 const HEIGHT: i32 = 400;