+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
 
-## [1.2.7] - Unreleased
+## [1.2.7] - 2021-10-08
 - Add Group::try_current().
+- Add Color::from_rgba_tuple().
+- Constify more enums::Color and Shortcut methods.
+- Constify utils functions.
+- Update FLTK.
 
 ## [1.2.6] - 2021-10-05
 - Move macros into fltk and remove dependency on syn & quote.